+++ packages/web/src/modules/profiles/components/layout/ProfileLayout.vue
@@ -1,10 +1,9 @@
 <script lang="ts" setup>
 import { Size } from '@/modules/ui/types';
-import ProfileSidebar from './snippets/ProfileSidebar.vue';
+import ProfileDrawer from './snippets/ProfileDrawer.vue';
 import MainProfileContainer from './MainProfileContainer.vue';
-import { computed, defineAsyncComponent } from 'vue';
+import { computed, defineAsyncComponent, Ref, ref } from 'vue';
 import { storeToRefs } from 'pinia';
-import { useAuthStore } from '@/modules/auth/store/auth.store';
 import CreateProfileModal from '../modals/CreateProfileModal.vue';
 import InviteUsersModal from '../../../invitations/components/modals/InviteUsersModal.vue';
 import { useIntroductionTourStore } from '@/modules/help/stores/introduction-tour.store';
@@ -13,6 +12,17 @@ import HelpModal from '@/modules/help/components/HelpModal.vue';
 import CreateOrEditContentModal from '@/modules/content-stream/components/CreateOrEditContentModal.vue';
 import MobileFotterMenu from '@/modules/profiles/components/menus/MobileFotterMenu.vue';
 import { usePageStore } from '@/modules/core/store/page.store';
+import { useSwipe } from '@vueuse/core';
+import { useAccountStore } from '@/modules/account/stores/account.store';
+
+const pageStore = usePageStore();
+const accountStore = useAccountStore();
+
+const { toggleSidebar } = pageStore;
+const { showAccountDrawer } = storeToRefs(accountStore);
+
+const { showMobileFooter, showSidebar } = storeToRefs(pageStore);
+const root = ref<HTMLElement>() as Ref<HTMLElement>;
 
 export interface IProps {
   containerWidth?: 'xs' | 'sm' | 'lg' | 'xl' | 'full';
@@ -32,9 +42,15 @@ const { active: showIntroductionTour } = storeToRefs(useIntroductionTourStore())
 
 const containerProps = computed(() => ({ width: props.containerWidth }));
 
-const show = computed(() => (props.requireAuth ? useAuthStore().isAuthenticated : true));
+const { direction } = useSwipe(root, {
+  onSwipeEnd(e: TouchEvent) {
+    // Closing swipe gestures are handled within the drawer components
+    if (showSidebar.value || showAccountDrawer.value) return;
 
-const { showMobileFooter } = storeToRefs(usePageStore());
+    if (direction.value === 'right') toggleSidebar();
+    if (direction.value === 'left') showAccountDrawer.value = true;
+  },
+});
 
 const IntroductionTour = defineAsyncComponent(
   () => import('@/modules/help/components/IntroductionTour.vue'),
@@ -42,8 +58,8 @@ const IntroductionTour = defineAsyncComponent(
 </script>
 
 <template>
-  <profile-sidebar v-if="show" />
-  <div v-if="show" class="overflow-hidden flex w-full h-screen-s flex-col">
+  <profile-drawer />
+  <div ref="root" class="overflow-hidden flex w-full h-screen-s flex-col">
     <div class="flex items-stretch flex-col h-screen-s">
       <main-profile-container v-bind="containerProps">
         <slot>