+++ packages/web/src/modules/ui/components/slider/SliderNavigation.vue
@@ -65,11 +65,11 @@ const sliderStyle = computed(() => {
 <template>
   <div
     ref="container"
-    class="no-swipe flex tag-filter-selection overflow-x-hidden whitespace-nowrap relative mr-1">
+    class="no-swipe flex tag-filter-selection overflow-x-hidden whitespace-nowrap relative">
     <div
       ref="slider"
       :style="sliderStyle"
-      class="inline-flex slider-nav touch-pan-y slider-nav flex"
+      class="inline-flex gap-0.5 pr-0.5 slider-nav touch-pan-y slider-nav flex"
       @mousedown="beginSlide"
       @touchstart="beginSlide">
       <slot></slot>