+++ CHANGELOG.md
@@ -5,15 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.0-alpha3] - WIP
+## [0.1.0-alpha4] - WIP
 
 ### Fixed
 - Swipe handler should be prevented in case of active scroll
+- MainContainer does not have full width and scrollbar is not at the right page
 
 ### Changed
 - Mobile profile drawer is closed on route change
 
-## [0.1.0-alpha3] - 2023-06-06
+## [0.1.0-alpha4] - 2023-06-06
 
 ### Fixed
 - Modal autofocus was not working due to new dialog